moved writer to own file
